backend/services/medgemma_colab.py
"""
MedGemma 4B via Google Colab - Privacy-First Medical AI
"""
import requests
import base64
from PIL import Image
import io
import os
from typing import Dict, Any

class MedGemmaColab:
    def __init__(self):
        self.colab_url = os.getenv('MEDGEMMA_COLAB_URL', '').rstrip('/')
        self.enabled = bool(self.colab_url)
    
    def analyze_xray(self, image: Image.Image, prompt: str = None) -> Dict[str, Any]:
        """
        Analyze X-ray image using self-hosted MedGemma 4B
        
        Args:
            image: PIL Image object
            prompt: Optional custom prompt
            
        Returns:
            Dict with analysis results
        """
        if not self.enabled: