backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
python-multipart
cryptography>=41.0.0
twilio
scikit-learn
numpy
PyPDF2
pytesseract
Pillow
pdf2image
requests
web3>=6.0.0
eth-utils>=2.0.0
pytest>=7.0.0
google-cloud-aiplatform>=1.38.0
reportlab>=4.0.0
gradio_client>=1.0.0
stellar-sdk>=9.1.0
